LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-complex-unresolved-serialize-expected.txt


FAIL 'calc(pow(2, sign(1em - 18px)))' as a specified value should serialize as 'calc(pow(2, sign(1em - 18px)))'. assert_equals: 'calc(pow(2, sign(1em - 18px)))' should round-trip exactly in specified values. expected "calc(pow(2, sign(1em - 18px)))" but got "pow(2, sign(1em - 18px))"
PASS 'calc(pow(2, sign(1em - 18px)))' as a computed value should serialize as '2'.
FAIL 'calc(pow(sign(1em - 18px), 2))' as a specified value should serialize as 'calc(pow(sign(1em - 18px), 2))'. assert_equals: 'calc(pow(sign(1em - 18px), 2))' should round-trip exactly in specified values. expected "calc(pow(sign(1em - 18px), 2))" but got "pow(sign(1em - 18px), 2)"
PASS 'calc(pow(sign(1em - 18px), 2))' as a computed value should serialize as '1'.
FAIL 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' as a specified value should serialize as 'calc(pow(sign(1em - 18px), sign(1em - 18px)))'. assert_equals: 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' should round-trip exactly in specified values. expected "calc(pow(sign(1em - 18px), sign(1em - 18px)))" but got "pow(sign(1em - 18px), sign(1em - 18px))"
PASS 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' as a computed value should serialize as '1'.
FAIL 'calc(pow(2, sibling-index())' as a specified value should serialize as 'calc(pow(2, sibling-index()))'. assert_not_equals: 'calc(pow(2, sibling-index())' should be valid in orphans. got disallowed value ""
FAIL 'calc(pow(2, sibling-index())' as a computed value should serialize as '2'. assert_equals: 'calc(pow(2, sibling-index())' and '2' should serialize the same in computed values. expected "2" but got "auto"
FAIL 'calc(pow(sibling-index(), 2)' as a specified value should serialize as 'calc(pow(sibling-index(), 2))'. assert_not_equals: 'calc(pow(sibling-index(), 2)' should be valid in orphans. got disallowed value ""
FAIL 'calc(pow(sibling-index(), 2)' as a computed value should serialize as '1'. assert_equals: 'calc(pow(sibling-index(), 2)' and '1' should serialize the same in computed values. expected "1" but got "auto"
FAIL 'calc(pow(sibling-index(), sibling-index())' as a specified value should serialize as 'calc(pow(sibling-index(), sibling-index()))'. assert_not_equals: 'calc(pow(sibling-index(), sibling-index())' should be valid in orphans. got disallowed value ""
FAIL 'calc(pow(sibling-index(), sibling-index())' as a computed value should serialize as '1'. assert_equals: 'calc(pow(sibling-index(), sibling-index())' and '1' should serialize the same in computed values. expected "1" but got "auto"
